root/backend/requirements.txt
Django==5.0.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
gunicorn==21.2.0
python-dotenv==1.0.1
Pillow==10.2.0  # For image handling
whitenoise==6.6.0  # For serving static files
